[accounts] /accounts/update/profile

+++ accounts.py
@@ -184,7 +184,8 @@ def update_profile():
 
             shutil.copy(file_path, random_file_path)
 
-            os.remove(user.full_body_img_path)
+            if user.full_body_img_path:
+                os.remove(user.full_body_img_path)
             user.full_body_img_path = random_file_path
 
         else:
